extracts error handling to seperate module
